+++ src/components/orderArea/ProductHole.jsx
@@ -1,9 +1,9 @@
 import React, { useContext } from 'react';
-import { Container, Product } from 'components/orderArea/ProductHoll.style';
+import { Container, Product } from 'components/orderArea/ProductHole.style';
 // import { FinalPayContext } from 'pages/VendingMachine';
 import { SelectedProductContext } from 'App';
 
-export default function ProductHoll() {
+export default function ProductHole() {
   const { detail } = useContext(SelectedProductContext)[0];
 
   return (
